# docs/swagger.yaml
basePath: /
definitions:
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.AddTaskRequest:
    properties:
      taskName:
        type: string
      userId:
        type: integer
    type: object
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.AddUserRequest:
    properties:
      passportNumber:
        type: string
    type: object
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.TaskResponse:
    properties:
      endTime:
        type: string
      startTime:
        type: string
      taskId:
        type: integer
      taskName:
        type: string
      userId:
        type: integer
    type: object
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.UpdateTaskRequest:
    properties:
      isFinished:
        type: boolean
      name:
        type: string
      taskId:
        type: integer
    type: object
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.UpdateUserRequest:
    properties:
      passportNumber:
        type: string
      userId:
        type: integer
    type: object
  github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.UserResponse:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      passportNumber:
        type: string
      patronymic:
        type: string
      surname:
        type: string
    type: object
externalDocs:
  description: OpenAPI 2.0
host: petstore.swagger.io
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the backend server for the test assignment.
  license:
    name: .
    url: .
  termsOfService: http://swagger.io/terms/
  title: Time Tracker
  version: "1.0"
paths:
  /tasks:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.AddTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task created successfully
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.SuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Add a new task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Updated task details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated successfully
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.TaskResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Update task details
      tags:
      - Tasks
  /tasks/{taskId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task deleted successfully
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.SuccessResponse'
        "400":
          description: Invalid task ID
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Delete task by Id
      tags:
      - Tasks
  /tasks/{userId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            items:
              $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.TaskResponse'
            type: array
        "400":
          description: Invalid user ID or pagination parameters
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Get tasks by user ID with pagination
      tags:
      - Tasks
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.AddUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User added successfully
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.SuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Add a new user
      tags:
      - Users
  /users/{userId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Delete user by Id
      tags:
      - Users
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.UserResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Get user by Id
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      - description: Updated user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.SuccessResponse'
        "400":
          description: Invalid request payload or user ID
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_XT3RM1NAT0R_time-tracker_internal_delivery_models.ErrorResponse'
      summary: Update user details
      tags:
      - Users
swagger: "2.0"